Modularize jssp example
