+++ examples/jssp/problem/individual.rs
@@ -1,178 +1,15 @@
 use std::collections::HashSet;
 
-use ecrs::{
-    ga::{self, Individual},
-    prelude::population::PopulationGenerator,
-};
+use crate::util::{print_slice, print_hash_set};
 
-use crate::util::{print_hash_set, print_slice};
+use super::{Operation, Machine};
 
 #[derive(Debug)]
 pub struct JsspIndividual {
     pub chromosome: Vec<f64>,
-    operations: Vec<Operation>,
-    machines: Vec<Machine>,
+    pub operations: Vec<Operation>,
+    pub machines: Vec<Machine>,
     pub fitness: usize,
-}
-
-#[derive(Debug)]
-pub struct Operation {
-    id: usize,
-    finish_time: usize,
-    duration: usize,
-    machine: usize,
-
-    // Should I hold references to other operations or just their ids
-    preds: Vec<usize>,
-}
-
-#[derive(Debug)]
-pub struct Machine {
-    id: usize,
-    // dummy structure
-    rmc: Vec<usize>,
-}
-
-impl Machine {
-    pub fn is_idle(&self, range: std::ops::RangeInclusive<usize>) -> bool {
-        assert!(*range.end() < 40);
-
-        for i in range {
-            if self.rmc[i] == 0 {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    pub fn reserve(&mut self, range: std::ops::Range<usize>) {
-        for i in range.clone() {
-            self.rmc[i] = 0;
-        }
-        println!("Reserved {}..{} in machine {}: ", range.start, range.end, self.id);
-        print_slice(&self.rmc);
-    }
-}
-
-#[derive(Debug)]
-pub struct JsspConfig {
-    pub n_jobs: usize,
-    pub n_machines: usize,
-}
-
-#[derive(Debug)]
-pub struct JsspState {
-    pub cfg: JsspConfig,
-    pub population: Vec<JsspIndividual>,
-}
-
-impl JsspState {
-    fn build_operations(&self) -> Vec<Operation> {
-        // For now I'm implementing just for single case
-        assert!(self.cfg.n_jobs == 4 && self.cfg.n_machines == 2);
-
-        let mut operations = Vec::with_capacity(self.cfg.n_jobs + 2);
-        operations.push(Operation {
-            id: 0,
-            finish_time: usize::MAX,
-            duration: 0,
-            machine: 0,
-            preds: Vec::new(),
-        });
-
-        operations.push(Operation {
-            id: 1,
-            finish_time: usize::MAX,
-            duration: 4,
-            machine: 1,
-            // preds: Vec::new(), // or maybe should I put Operation 0 here?
-            preds: vec![0],
-        });
-
-        operations.push(Operation {
-            id: 2,
-            finish_time: usize::MAX,
-            duration: 2,
-            machine: 0,
-            preds: vec![0, 1],
-        });
-
-        operations.push(Operation {
-            id: 3,
-            finish_time: usize::MAX,
-            duration: 1,
-            machine: 0,
-            // preds: Vec::new(),
-            preds: vec![0],
-        });
-
-        operations.push(Operation {
-            id: 4,
-            finish_time: usize::MAX,
-            duration: 3,
-            machine: 1,
-            preds: vec![0, 3],
-        });
-
-        operations.push(Operation {
-            id: 5,
-            finish_time: usize::MAX,
-            duration: 0,
-            machine: 0,
-            preds: vec![0, 1, 2, 3, 4],
-        });
-
-        operations
-    }
-
-    fn build_machines(&self) -> Vec<Machine> {
-        // To represent machine we could think of some sparse structure, but for now let it be
-        // vector
-
-        let machines = vec![
-            Machine {
-                id: 0,
-                rmc: vec![1; 40],
-            },
-            Machine {
-                id: 1,
-                rmc: vec![1; 40],
-            },
-        ];
-
-        machines
-    }
-
-    pub fn init_pop(&mut self, size: usize) {
-        self.population = ga::population::RandomPoints::with_single_constraint(8, 0.0..1.0)
-            .generate(size)
-            .into_iter()
-            .map(|idv| JsspIndividual {
-                chromosome: idv.chromosome,
-                operations: self.build_operations(),
-                fitness: usize::MAX,
-                machines: self.build_machines(),
-            })
-            .collect();
-    }
-
-    pub fn inject_ecrs_pop(&mut self, population: Vec<Individual<Vec<f64>>>) {
-        self.population = population
-            .into_iter()
-            .map(|idv| JsspIndividual {
-                chromosome: idv.chromosome,
-                operations: self.build_operations(),
-                fitness: usize::MAX,
-                machines: self.build_machines(),
-            })
-            .collect();
-    }
-
-    pub fn eval_pop(&mut self) {
-        for idv in self.population.iter_mut() {
-            idv.eval();
-        }
-    }
 }
 
 impl JsspIndividual {
@@ -209,7 +46,7 @@ impl JsspIndividual {
             })
     }
 
-    fn eval(&mut self) -> usize {
+    pub fn eval(&mut self) -> usize {
         println!("++++++++++++++++++++++++++++++++++");
         // We deduce the problem size from the chromosome size
         let n: usize = self.chromosome.len() / 2;
